adjust tests - add syncTables script - add more tests
