--- server/models/Article.js
@@ -1,30 +1,0 @@
-import { DataTypes } from 'sequelize';
-import db from '../config/db';
-import User from './User';
-
-const Article = db.define('article', {
-  authorId: {
-    type: DataTypes.INTEGER,
-    allowNull: false,
-    references: {
-      model: User,
-      key: 'id',
-    },
-  },
-  authorName: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-  title: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-  article: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-}, {
-  timestamps: true,
-  updatedAt: false,
-});
-export default Article;
